requirements.txt
aiohttp==3.8.1
asgiref==3.5.2
cn2an==0.5.19
Django==4.1
djangorestframework==3.13.1
django-environ==0.9.0
django-filter==22.1
lxml==4.9.2
qbittorrent-api==2022.8.36
requests==2.28.1
tzlocal==4.2
whitenoise==6.2.0
apscheduler==3.9.1
psutil==5.9.2
packaging==23.0